requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioice==0.10.1
aiortc==1.11.0
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
asyncer==0.0.7
attrs==25.3.0
av==14.3.0
azure-core==1.33.0
azure-identity==1.21.0
backoff==2.2.1
bidict==0.23.1
certifi==2025.4.26
cffi==1.17.1
chainlit==2.5.5
chardet==5.2.0
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
cloudevents==1.11.0
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
frozenlist==1.6.0
fsspec==2025.3.2
google-crc32c==1.7.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.6.1
inflection==0.5.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
Lazify==0.4.0
lazy-object-proxy==1.11.0
literalai==0.1.201
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.6.0
monotonic==1.6
more-itertools==10.7.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.2.5
openai==1.76.0
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.39.4
opentelemetry-instrumentation-anthropic==0.39.4
opentelemetry-instrumentation-bedrock==0.39.4
opentelemetry-instrumentation-chromadb==0.39.4
opentelemetry-instrumentation-cohere==0.39.4
opentelemetry-instrumentation-crewai==0.39.4
opentelemetry-instrumentation-google-generativeai==0.39.4
opentelemetry-instrumentation-groq==0.39.4
opentelemetry-instrumentation-haystack==0.39.4
opentelemetry-instrumentation-lancedb==0.39.4
opentelemetry-instrumentation-langchain==0.39.4
opentelemetry-instrumentation-llamaindex==0.39.4
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.39.4
opentelemetry-instrumentation-milvus==0.39.4
opentelemetry-instrumentation-mistralai==0.39.4
opentelemetry-instrumentation-ollama==0.39.4
opentelemetry-instrumentation-openai==0.39.4
opentelemetry-instrumentation-pinecone==0.39.4
opentelemetry-instrumentation-qdrant==0.39.4
opentelemetry-instrumentation-replicate==0.39.4
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.39.4
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.39.4
opentelemetry-instrumentation-transformers==0.39.4
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.39.4
opentelemetry-instrumentation-watsonx==0.39.4
opentelemetry-instrumentation-weaviate==0.39.4
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.3
opentelemetry-util-http==0.52b1
packaging==25.0
parse==1.20.2
pathable==0.4.4
posthog==3.25.0
prance==23.6.21.0
propcache==0.3.1
protobuf==5.29.4
pybars4==0.9.13
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pyee==13.0.0
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMeta3==0.5.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-multipart==0.0.18
python-socketio==5.13.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.2
semantic-kernel==1.28.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.3
starlette==0.41.3
syncer==2.0.3
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
traceloop-sdk==0.39.4
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uptrace==1.31.0
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==0.20.0
websockets==15.0.1
Werkzeug==3.1.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.21.0
